Refactor major structures and flow
This change has a number of significant, breaking modifications:

* Went from private internal commands to public Register
* Changed and clarified how Call() works
* Clarified New() and what the arguments mean
* Drop automatic inclusion of builtins (temporarily)

This commit should not be considered finished since builtins are moving
to `prefab` and `errors.go` needs an overhaul and migration to its own
package as well.

Please not we have *not* hit even v0.1 yet and this is still a very
volatile working branch despite it being on main.